# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderly Trails Travel Website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Destination {
  id: number;
  name: string;
  country: string;
  description: string;
  imageUrl: string;
  category: string;
  startingPrice: number;
  rating: number;
  bestSeason: string;
  /** @nullable */
  weather?: string | null;
  featured?: boolean;
}

export interface Package {
  id: number;
  title: string;
  destinationId: number;
  destinationName: string;
  imageUrl: string;
  price: number;
  duration: number;
  nights: number;
  category: string;
  rating: number;
  hotelStars: number;
  mealsIncluded: boolean;
  transportIncluded: boolean;
  description: string;
  /** @nullable */
  itinerary?: string | null;
  /** @nullable */
  includedItems?: string | null;
  /** @nullable */
  excludedItems?: string | null;
  featured: boolean;
}

export interface PackageInput {
  title: string;
  destinationId: number;
  imageUrl?: string;
  price: number;
  duration: number;
  nights: number;
  category: string;
  hotelStars: number;
  mealsIncluded: boolean;
  transportIncluded: boolean;
  description: string;
  itinerary?: string;
  includedItems?: string;
  excludedItems?: string;
  featured?: boolean;
}

export type PackageStatsCategoryCountsItem = {
  category: string;
  count: number;
};

export interface PackageStats {
  totalPackages: number;
  totalDestinations: number;
  avgPrice: number;
  totalBookings: number;
  categoryCounts: PackageStatsCategoryCountsItem[];
}

export interface Booking {
  id: number;
  name: string;
  email: string;
  phone: string;
  destination: string;
  /** @nullable */
  packageId?: number | null;
  travelers: number;
  budget: string;
  travelDate: string;
  /** @nullable */
  specialRequests?: string | null;
  status: string;
  createdAt: string;
}

export interface BookingInput {
  name: string;
  email: string;
  phone: string;
  destination: string;
  packageId?: number;
  travelers: number;
  budget: string;
  travelDate: string;
  specialRequests?: string;
}

export interface BlogPost {
  id: number;
  title: string;
  excerpt: string;
  content: string;
  imageUrl: string;
  category: string;
  author: string;
  publishedAt: string;
  readTime: number;
}

export interface BlogPostInput {
  title: string;
  excerpt: string;
  content: string;
  imageUrl: string;
  category: string;
  author: string;
  readTime: number;
}

export interface Testimonial {
  id: number;
  name: string;
  location: string;
  rating: number;
  review: string;
  avatarUrl: string;
  packageName: string;
}

export interface TestimonialInput {
  name: string;
  location: string;
  rating: number;
  review: string;
  avatarUrl: string;
  packageName: string;
}

export interface NewsletterInput {
  email: string;
}

export interface NewsletterSubscription {
  id: number;
  email: string;
  subscribedAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone: string;
  subject?: string;
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  subject?: string | null;
  message: string;
  createdAt: string;
}

export type ListDestinationsParams = {
  category?: string;
};

export type ListPackagesParams = {
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  duration?: number;
};

export type ListBlogPostsParams = {
  category?: string;
};