# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderly Trails Travel Website API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  name: string;
  email: string;
  phone: string;
  destination: string;
  /** @nullable */
  packageId?: number | null;
  travelers: number;
  budget: string;
  travelDate: string;
  /** @nullable */
  specialRequests?: string | null;
  status: string;
  createdAt: string;
}