# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderly Trails Travel Website API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  name: string;
  email: string;
  phone: string;
  destination: string;
  packageId?: number;
  travelers: number;
  budget: string;
  travelDate: string;
  specialRequests?: string;
}