# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderly Trails Travel Website API
 * OpenAPI spec version: 0.1.0
 */

export interface Destination {
  id: number;
  name: string;
  country: string;
  description: string;
  imageUrl: string;
  category: string;
  startingPrice: number;
  rating: number;
  bestSeason: string;
  /** @nullable */
  weather?: string | null;
  featured?: boolean;
}