# lib/api-zod/src/generated/types/newsletterSubscription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderly Trails Travel Website API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterSubscription {
  id: number;
  email: string;
  subscribedAt: string;
}