# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderly Trails Travel Website API
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  id: number;
  title: string;
  destinationId: number;
  destinationName: string;
  imageUrl: string;
  price: number;
  duration: number;
  nights: number;
  category: string;
  rating: number;
  hotelStars: number;
  mealsIncluded: boolean;
  transportIncluded: boolean;
  description: string;
  /** @nullable */
  itinerary?: string | null;
  /** @nullable */
  includedItems?: string | null;
  /** @nullable */
  excludedItems?: string | null;
  featured: boolean;
}