# lib/api-zod/src/generated/types/packageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderly Trails Travel Website API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageInput {
  title: string;
  destinationId: number;
  imageUrl?: string;
  price: number;
  duration: number;
  nights: number;
  category: string;
  hotelStars: number;
  mealsIncluded: boolean;
  transportIncluded: boolean;
  description: string;
  itinerary?: string;
  includedItems?: string;
  excludedItems?: string;
  featured?: boolean;
}